Not do delete stale tc rules when no filter/qdisc existed
If there has a floating IP with rate_limit 0, and this floating
IP is the first one in the DVR namespace. Then there have no qdisc
or filter rules in that namespace. Then _delete_stale_tc_rules will
get a iterable TypeError.

This patch will check that get_existed_filter_ids return value.

+++ neutron/agent/l3/router_info.py
@@ -922,10 +922,10 @@ class RouterInfo(object):
                         new_filters.add(filter_id)
                 except tc_lib.FilterIDForIPNotFound:
                     pass
-            existed_filter_ids = set(
-                tc_wrapper.get_existed_filter_ids(direction))
-            removed_filters = existed_filter_ids - new_filters
-            tc_wrapper.delete_filter_ids(direction, removed_filters)
+            existed_filter_ids = tc_wrapper.get_existed_filter_ids(direction)
+            if existed_filter_ids:
+                removed_filters = set(existed_filter_ids) - new_filters
+                tc_wrapper.delete_filter_ids(direction, removed_filters)
 
     def process_external(self, agent):
         fip_statuses = {}
